+++ ClickerArch/Assets/Services/EnemyServices/IEnemyView.cs
@@ -7,7 +7,7 @@ public interface IEnemyView
     void Idle();
     void Attack();
     void Death();
-    void Hurt(float ratio, int damage, bool showDamage);
+    void Hurt(float ratio, double damage, bool showDamage);
 
     void ConfigureForId(string id);
 }
